old-tests/submission/W3C/domhtml/HTMLObjectElement18.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - OBJECT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/the-object-element.html#dom-object-name" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<object name='object1' data="/support/html/sample.html" height="80" type="text/html" width="550"></object>

<script>
var domtest = async_test("The IDL attribute name must reflect the respective content attributes of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vname;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("object");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vname = testNode.name;

      assert_equals(vname,"object1");
       
    domtest.done();
}));

</script>
<p>The IDL attribute name must reflect the respective content attributes of the same name.</p>
<div id="log"></div>
</body></html>
